requirements.txt
flask
python-dotenv
flask-cors
gunicorn
flask-sqlalchemy